pyproject.toml
[project]
name = "ai-devs-builders"
version = "0.1.0"
description = "AI development training exercises and supporting tools."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "requests>=2.33.1",
    "loguru>=0.7.3",
    "langchain>=1.2.15",
    "langfuse>=4.2.0",
    "streamlit>=1.56.0",
    "python-dotenv>=1.2.2",
    "beautifulsoup4>=4.14.3",
    "openai>=2.31.0",
    "faster-whisper>=1.2.1",
    "typing>=3.10.0.0",
    "edge-tts>=7.2.8",
    "asyncio>=4.0.0",
    "pathlib>=1.0.1",
    "argparse>=1.4.0",
    "playwright>=1.58.0",
    "pydantic>=2.13.0",
    "fastapi>=0.135.3",
    "langchain-openai>=1.2.1",
    "rapidfuzz>=3.14.5",
]
